+++ src/bosh-alicloud-cpi/alicloud/common.go
@@ -83,7 +83,8 @@ const (
 func getSdkConfig() *sdk.Config {
 	return sdk.NewConfig().
 		WithMaxRetryTime(5).
-		WithUserAgent("Bosh-Alicloud-Cpi").
+		WithTimeout(time.Duration(60) * time.Second).
+		WithUserAgent(getUserAgent()).
 		WithGoRoutinePoolSize(10).
 		WithDebug(false).
 		WithHttpTransport(getTransport())
@@ -96,4 +97,8 @@ func getTransport() *http.Transport {
 	}
 	return &http.Transport{
 		TLSHandshakeTimeout: time.Duration(handshakeTimeout) * time.Second}
+}
+
+func getUserAgent() string {
+	return "Bosh-Alicloud-Cpi"
 }